+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * `rcu` support.
 * `compare_and_swap` support.
+* Added some primitive benchmarks.
 
 # 0.1.0